Flip Height and Width Args on CreateBitmap & bitmap->Create
Changelog: Flipped thew order of int height, int width for Game->CreateBitmap()
and for bitmap->Create().

Added a QR to use the old args order, set for quests with a build ID < 44.
Enabled that QR for older quests on quest load.
Fixed some sign mismatches.

+++ src/ffscript.h
@@ -1022,7 +1022,7 @@ enum __Error
         {
             long gptr = ptr - MAX_ZCARRAY_SIZE;
             
-            if(gptr > game->globalRAM.size())
+            if(gptr > (long)game->globalRAM.size())
                 return InvalidError(ptr);
                 
             return game->globalRAM[gptr];
@@ -1101,7 +1101,7 @@ enum __Error
         
         if(a == INVALIDARRAY)
             return;
-	if ( getSize(ptr) < min_size )
+	if ( getSize(ptr) < (unsigned)min_size )
 	{
 		Z_scripterrlog("Invalid array, size of [%d] provided to a function requiring a size of [%d].\n", getSize(ptr), min_size);
         }
